Transformations.Relabeler.Relabeler

AST Relabeler

relabelAst :: Program -> Program

Relabel identifiers in the AST so that each declaration creates an identifier with a unique name.

It helps to avoid naming errors in the future.

Internal

Relabeler State

type RelabelerState = State Env

data Env

Constructors

Env

Fields

idMappings :: [IdMapping]

idCnt :: IdCnt

type IdMapping = (Identifier', Identifier')

Relabelers

relabelDecl :: Declaration -> RelabelerState Declaration

relabelExpr :: Expression -> RelabelerState Expression

relabelFun :: Fun -> RelabelerState Fun

Identifier Mappings

mapId :: Identifier' -> RelabelerState Identifier'

pushAndMapId :: Identifier' -> RelabelerState Identifier'

pushId :: Identifier' -> RelabelerState ()

popId :: RelabelerState ()

Utils

relabel1 :: (Expression -> a) -> Expression -> RelabelerState a

relabel2 :: (Expression -> Expression -> a) -> Expression -> Expression -> RelabelerState a

relabel3 :: (Expression -> Expression -> Expression -> a) -> Expression -> Expression -> Expression -> RelabelerState a