TypeChecker.HindleyMilner

Type

data TypeF a

Constructors

TVarF Identifier

TUnitF

TBoolF

TIntF

TFunF a a

UType

type UType = UTerm TypeF IntVar

data UTerm (t :: Type -> Type) v where

The type of terms generated by structures t over variables v. The structure type should implement Unifiable and the variable type should implement Variable.

The Show instance doesn't show the constructors, in order to improve legibility for large terms.

All the category theoretic instances (Functor, Foldable, Traversable,...) are provided because they are often useful; however, beware that since the implementations must be pure, they cannot read variables bound in the current context and therefore can create incoherent results. Therefore, you should apply the current bindings before using any of the functions provided by those classes.

Bundled Patterns

pattern UTVar :: Identifier -> UType

pattern UTUnit :: UType

pattern UTBool :: UType

pattern UTInt :: UType

pattern UTFun :: UType -> UType -> UType

data UTerm (t :: Type -> Type) v where

The type of terms generated by structures t over variables v. The structure type should implement Unifiable and the variable type should implement Variable.

The Show instance doesn't show the constructors, in order to improve legibility for large terms.

All the category theoretic instances (Functor, Foldable, Traversable,...) are provided because they are often useful; however, beware that since the implementations must be pure, they cannot read variables bound in the current context and therefore can create incoherent results. Therefore, you should apply the current bindings before using any of the functions provided by those classes.

Bundled Patterns

pattern UTVar :: Identifier -> UType

pattern UTUnit :: UType

pattern UTBool :: UType

pattern UTInt :: UType

pattern UTFun :: UType -> UType -> UType

data UTerm (t :: Type -> Type) v where

The type of terms generated by structures t over variables v. The structure type should implement Unifiable and the variable type should implement Variable.

The Show instance doesn't show the constructors, in order to improve legibility for large terms.

All the category theoretic instances (Functor, Foldable, Traversable,...) are provided because they are often useful; however, beware that since the implementations must be pure, they cannot read variables bound in the current context and therefore can create incoherent results. Therefore, you should apply the current bindings before using any of the functions provided by those classes.

Bundled Patterns

pattern UTVar :: Identifier -> UType

pattern UTUnit :: UType

pattern UTBool :: UType

pattern UTInt :: UType

pattern UTFun :: UType -> UType -> UType

data UTerm (t :: Type -> Type) v where

The type of terms generated by structures t over variables v. The structure type should implement Unifiable and the variable type should implement Variable.

The Show instance doesn't show the constructors, in order to improve legibility for large terms.

All the category theoretic instances (Functor, Foldable, Traversable,...) are provided because they are often useful; however, beware that since the implementations must be pure, they cannot read variables bound in the current context and therefore can create incoherent results. Therefore, you should apply the current bindings before using any of the functions provided by those classes.

Bundled Patterns

pattern UTVar :: Identifier -> UType

pattern UTUnit :: UType

pattern UTBool :: UType

pattern UTInt :: UType

pattern UTFun :: UType -> UType -> UType

data UTerm (t :: Type -> Type) v where

The type of terms generated by structures t over variables v. The structure type should implement Unifiable and the variable type should implement Variable.

The Show instance doesn't show the constructors, in order to improve legibility for large terms.

All the category theoretic instances (Functor, Foldable, Traversable,...) are provided because they are often useful; however, beware that since the implementations must be pure, they cannot read variables bound in the current context and therefore can create incoherent results. Therefore, you should apply the current bindings before using any of the functions provided by those classes.

Bundled Patterns

pattern UTVar :: Identifier -> UType

pattern UTUnit :: UType

pattern UTBool :: UType

pattern UTInt :: UType

pattern UTFun :: UType -> UType -> UType

Polytype

data Poly t

Constructors

Forall [Identifier] t

type Polytype = Poly Type

Converters

toUType :: Type -> UType

toPolytype :: UPolytype -> Polytype

Infer

type Infer = ReaderT Ctx (ExceptT TypeError (IntBindingT TypeF Identity))

lookup :: Identifier -> Infer UType

withBinding :: MonadReader Ctx m => Identifier -> UPolytype -> m a -> m a

FreeVars

fresh :: Infer UType

Errors

data TypeError where

Constructors

Unreachable :: TypeError

UnboundVar :: Identifier -> TypeError

Infinite :: IntVar -> UType -> TypeError

ImpossibleBinOpApplication :: UType -> UType -> TypeError

ImpossibleUnOpApplication :: UType -> TypeError

Mismatch :: TypeF UType -> TypeF UType -> TypeError

(=:=) :: UType -> UType -> Infer UType

applyBindings :: UType -> Infer UType

mkVarName :: String -> IntVar -> Identifier

generalize :: UType -> Infer UPolytype

Orphan instances

Ord IntVar

Instance details

Methods

compare :: IntVar -> IntVar -> Ordering

(<) :: IntVar -> IntVar -> Bool

(<=) :: IntVar -> IntVar -> Bool

(>) :: IntVar -> IntVar -> Bool

(>=) :: IntVar -> IntVar -> Bool

max :: IntVar -> IntVar -> IntVar

min :: IntVar -> IntVar -> IntVar